+++ web-application-server-gradle/src/main/java/customwebserver/http/RequestLine.java
@@ -1,4 +1,4 @@
-package utils.dto;
+package customwebserver.http;
 
 import java.util.Map;
 import java.util.Objects;
